eunit.engine: first version of assertEqualFiles
This is the first version of the file comparison assertion, using the
Eclipse differencer in the Compare component. However, for some reason,
I have to bring SWT just because ITypedElement has Image as the return
type of getImage(). Once I get the first version done, I need to think
of a way to perform file/directory comparison without depending on the
UI. Ideally, it should work like for strings: no UI needed for comparing
them, only for showing the differences.