AbstractEmfModel: take snapshots of resource sets for model comparisons
Before switching to resource sets for comparisons, we were taking snapshots of
the EMF resource before doing the comparison. This was useful for producing
accurate results and working with models which were entirely in-memory or were
not going to be saved after the test.

This commit implements the same sort of functionality, but for resource sets.

Taking snapshots of a resource set before doing a comparison has an additional
and unexpected benefit: it works around the cache in the EmfModelResourceFactory.
This factory caches Resource objects by URI: it is beneficial most of the time,
but it can have unexpected side effects when comparing complex models.

For instance, a GMFGen model references several internal GMF models, such as
notation.ecore. Suppose we had two GMFGen models, each inside its own
set. Due to the caching in EmfModelResourceFactory, resolving the cross-reference
in the two GMFGen models to notation.ecore would produce the same Resource. As
EMF ensures that a Resource is only part of a single resource set, only one of
the resource sets would contain notation.ecore, and EMF Compare would report
differences between them.

Removing the cache altogether in EmfModelResourceFactory did work, but after
discussing with Dimitris, we decided to avoid this approach, as it could break
existing code written by the users.

In the end, it seems it is enough to save all non-platform models in both
resource sets to temporary files, and use regular EMF APIs to load them
without the cache in the EmfModelResourceFactory. We must not save platform
models to temporary files: this will make their filenames differ in the left
and right resource sets during comparison, and that tends to confuse EMF
Compare.