[EUnit UI] Fix bug in which the first launch would not be added to the history
The EUnit view normally uses a LaunchListener to keep track of all the launches
done and keep the history up to date. This LaunchListener is set up in the start()
method of the EUnitPlugin activator. This is a problem if the user hasn't loaded
the EUnit view *before* running the EUnit task: since the listener will be added
after the launch has been added (while the Ant task opens up the EUnit view),
the EUnit view wouldn't be notified about the first launch.

In that case, when trying to rerun the first launch, the user would get an incorrect
error message, and they'd have to run the tests once more to get the correct
behaviour.

This commit fixes the problem by calling launchesAdded manually with the existing
launches inside the start() method.