[Eugenia] Support GMF figure templates
In the GmfMap -> GmfGen step, some bits of Java code are generated and
embedded in the GmfGen model. To customize this code, we need to use
GMF figure templates.

This commit adds support to Eugenia for GMF figure templates. Assume
that the metamodel is in MyProject/model/a.emf. If the MyProject/templates-gmfmap
folder exists, it will be searched for custom GMF figure templates.