[EOL] Pull up wipeCaches() from EUnitModule into EolModule
Manual JUnit tests which do not use EUnit may also need to clear all cached
information in an EOL module to rerun it with a different set of tests.

In that case, they will be probably using a regular EolModule instead of an
EUnitModule. To reuse the cache wiping code between these two classes, this
commit pulls up the wipeCaches() method and renames it to clearCache(), like
the methods in EolOperation, EolFormalParameterList and EolFormalParameter.