Fix some failed tests in Windows + Indigo
We weren't calling Project.init() in the workflow tests. It seems Ant 1.7 was fine with it, but Indigo has Ant 1.8.x and it does not like that very much.

For some reason, properties in a JavaModel are sorted differently, but this was not important to our test, so I relaxed it a bit.

We still have some failing tests which are related to EGL, but they seem to be mostly due to differing line terminations in Windows (CRLF) and Unix derivatives (LF).