cmp.emf: add long common prefix to temporary clones and check for unmatched models
EMF Compare uses URI similarity metrics to match resources from one resource set to
another. In addition, it requires that the URIs are 70% similar to match them.

This poses a problem when we create the temporary clones, as we add a unique suffix
to them to create the temporary file. To counteract this, this commit adds a common
prefix to all those temporary clones. This fixes the Eugenia regression tests, but
EMF Compare's URI similarity indices might still pose a problem for end users whose
models have very different filenames and paths.