[EVL] Avoid NPE in addMarkers if validate had not been called
Explanation from Andrey Skorikov in the forum thread at [1]:

"Hi guys,

there is a bug in org.eclipse.epsilon.evl.emf.validation.CompositeEValidator.java at line 44:

if one of the validators fails none of the remaining validators are invoked
because the corresponding expression is not evaluated anymore. This causes a
NullPointerException in org.eclipse.epsilon.evl.emf.validation.EvlValidator.java
at line 160 when validating the children since the validation results are not
initialized.

Just replace the "short-circuit" AND operator (&&) with the normal one (&) to fix it.

Best regards,
Andrey Skorikov"

After looking at the code, I don't think the short-circuit logic is wrong. It's just
that we should protect the code adequately from producing a NullPointerException. For
this reason, it's better to ensure that the 'results' field is never null. This also
reduces the number of objects that need to be created.

[1]: http://www.eclipse.org/forums/index.php/mv/msg/321400/834757/#msg_834757