ModeLink: clean up and add support for the 'Force Exeed' option
The Exeed editor has several advantages over some of the existing EMF-based
model editors: it has drag-and-drop model weaving (using custom weaving
metamodels) and it can show unrooted objects in an EMF resource.

For this reason, it might be useful to force ModeLink to open a model with
the Exeed editor instead of the default editor. Usually the Exeed editors
were always used, but this was reversed in a previous commit. This commit
does not intend to revert that change, but rather provide an option for the
user.