[EGL] Fix initialisation of default output root for #404235
When the user did not call setOutputRoot or setRoot on an EglTemplateFactory then generating files could file with a cryptic error: "java.lang.IllegalArgumentException: URI has an
authority component". This was because EglPersistentTemplate#resolveFile had no base directory to resolve against.

I noticed that EglModule sensibly initialises the root to the same directory as the template that is being executed using EglTemplateFactory#initialiseRoot. (NB: initialiseRoot does nothing  when the user has already called #setRoot). To fix this bug and make the EglTemplateFactory API consistent with the EglTemplateFactoryModuleAdapter API, I've replicated this root initialisation behaviour for EglTemplateFactory#load.