[SQUASH] Self-contained test suite for surefire
This is a major change (merging from smadani/unittests)
to address the issue of not being able to run tests
from Maven, because many tests relied on being run in the IDE.
The main limitation has been uses of class.getResource()
rather than an agnostic way (e.g. when in a JAR file)
to load files and their dependencies. Consequently it's been
a massive PITA to get this working, especially given
how some tests are so reliant on Eclipse / the workspace
for resources and metamodel registration etc.
Despite best efforts, anything involving HUTN is output
of the question for now. The current state of this commit
also won't work with some EUnit tests.
However a good chunk of the test suite is functional.
Also credit goes to hhoyos for initial investigation
and ideas on this front.