Flexmi: allow overriding template attributes
When a Flexmi template has (expression) attributes defined as a
convention or best effort, but that sometimes it is necessary to
override them (e.g. a placeholder name for a list of certain type),
there was an issue with the mapping between element attributes and
model features. This commit fixes that.