Update target platform to Java 17 and Eclipse 2023-12
The JSON tests had to be updated to Jetty 12, which uses SLF4J 2.x, so
we also had to make sure that they do not explicitly depend on SLF4J as
it was causing plugged-in tests to fail with a message saying that the
Epsilon test bundle could not be found.

In order to find the cause for the above error, I had to change the
configuration of tycho-surefire-plugin in the "plugged" profile in
org.eclipse.epsilon.test/pom.xml to include this:

  <defaultStartLevel>
    <level>7</level>
    <autoStart>true</autoStart>
  </defaultStartLevel>

This forced all bundles to be started, which produced a helpful error
message saying that the org.eclipse.epsilon.test bundle could not be
started as it was getting two versions of SLF4J from different
dependency chains: chain 1 was a direct dependency from the JSON test
plugin to org.slf4j 1.7.x, and chain 2 was an indirect dependency
through Jetty 12.