Build javadocs only from plugins
The javadoc:aggregate task was failing due to Javadoc's inability to
find some of the test dependencies in the test plugins. Given that it
does not make much sense to have Javadocs for our test plugins, I have
limited the javadoc:aggregate task to the projects inside the plugins/
folder.