eol.dap: refactor variable references
The existing code would only work for simple variables, but it would
have required having a getReference(...) method for every type of
reference that we would have.

This commit introduces a generic interface (IVariableReference) for
anything that we can reference through a DAP integer, which we could
then use to inspect model element properties, the elements of a
collection, and so on.