eol.dap: add inspection of collections
This commit allows for inspecting Java collections. For small
collections (below the threshold in SuspendedState, currently 200
elements) we will list each element on its own. For larger collections,
we will initially only show slices (according to the constant in
SuspendedState, currently 100-element long), and then let the user
specify the slice they are interested in.

For slices, there is a minor optimisation where we will use List.subList
instead of creating our own list by iterating to the starting index.